requirements.txt
Flask
gradio_client
transformers
psycopg2-binary
python-dotenv
Werkzeug
